new files for expansion pack button and expansion pack model
