+++ lib/achievements_page/achievements_page_lists.dart
@@ -1,4 +1,4 @@
-import 'package:achievements/achievements_page/filter_menu.dart';
+import 'package:achievements/achievements_page/filter_menu/filter_menu.dart';
 import 'package:achievements/achievements_page/wish_list_builder.dart';
 import 'package:flutter/material.dart';
