+++ lib/achievements_page/filter_menu/expansion_pack_model.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class ExpansionPackModel {
+  final Image image;
+  final String name;
+  final String key;
+
+  ExpansionPackModel(
+      {required this.image, required this.name, required this.key});
+}
