+++ lib/achievements_page/filter_menu/filter_menu.dart
@@ -1,4 +1,6 @@
 import 'package:achievements/achievements_page/achievements_page_lists.dart';
+import 'package:achievements/achievements_page/filter_menu/expansion_pack_button.dart';
+import 'package:achievements/achievements_page/filter_menu/expansion_pack_model.dart';
 import 'package:achievements/achievements_page/wish_list_builder.dart';
 import 'package:flutter/material.dart';
 
@@ -14,6 +16,10 @@ class FilterMenu extends StatefulWidget {
 
 class _FilterMenuState extends State<FilterMenu> {
   final List<ExpansionPackModel> _toggledPacks = [];
+
+  void _closeEndDrawer() {
+    Navigator.of(context).pop();
+  }
 
   void _filterWishes(String key) {
     for (var wish in AchievementPageLists.wishes) {
@@ -70,12 +76,21 @@ class _FilterMenuState extends State<FilterMenu> {
                 FilledButton(
                   onPressed: () {
                     _startFilter();
+                    _closeEndDrawer();
                   },
                   child: const Text('Filter wishes'),
                 ),
                 FilledButton(
                   onPressed: () {
                     _toggledPacks.clear();
+                    widget.filteringList.addAll(AchievementPageLists.wishes);
+
+                    if (widget.onFilterButtonTap != null) {
+                      widget.onFilterButtonTap!();
+                    }
+
+                    _closeEndDrawer();
+
                     setState(() {});
                   },
                   child: const Text('Reset filters'),
@@ -90,6 +105,10 @@ class _FilterMenuState extends State<FilterMenu> {
 
   void _startFilter() {
     widget.filteringList.clear();
+    if(_toggledPacks.isEmpty) {
+      widget.filteringList.addAll(AchievementPageLists.wishes);
+      setState(() {});
+    }
 
     for (var pack in _toggledPacks) {
       _filterWishes(pack.key);
@@ -110,87 +129,4 @@ class _FilterMenuState extends State<FilterMenu> {
     print('filterMenu disposed');
     super.dispose();
   }
-}
-
-class ExpansionPackButton extends StatelessWidget {
-  ExpansionPackButton({
-    super.key,
-    required this.pack,
-    this.onTap,
-    this.isToggled,
-  });
-
-  final ExpansionPackModel pack;
-  final Function? onTap;
-  bool? isToggled;
-
-  WidgetStateProperty<CircleBorder>? _selectBorder() {
-    if(isToggled != null) {
-      if(isToggled!) {
-        return WidgetStateProperty.all(
-          CircleBorder(
-            side: BorderSide(
-              width: 2.0,
-              color: Colors.deepPurple[500]!,
-            ),
-          ),
-        );
-      }
-    }
-
-    return null;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-
-    return SizedBox(
-      width: 83,
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.start,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          FilledButton.icon(
-            onPressed: () {
-              if(onTap != null) {
-                onTap!();
-              }
-            },
-            clipBehavior: Clip.hardEdge,
-            label: pack.image,
-            style: ButtonStyle(
-                fixedSize: WidgetStateProperty.all(
-                  const Size.fromRadius(35),
-                ),
-                overlayColor: WidgetStateProperty.all(Colors.black12),
-                backgroundColor: WidgetStateProperty.all(Colors.transparent),
-                padding: WidgetStateProperty.all(EdgeInsets.zero),
-                shape: _selectBorder(),
-            ),
-          ),
-          const SizedBox(
-            height: 3,
-          ),
-          Text(
-            maxLines: 2,
-            pack.name,
-            style: TextStyle(
-                fontSize: 13,
-                fontWeight: FontWeight.w500,
-                color: Colors.grey[700]),
-            textAlign: TextAlign.center,
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class ExpansionPackModel {
-  final Image image;
-  final String name;
-  final String key;
-
-  ExpansionPackModel(
-      {required this.image, required this.name, required this.key});
 }
